Renameing poolConfig to poolConfigs to stay in line with tradingLimits
