+++ scripts/poolConfigs/README.md
