+++ scripts/poolConfigs/index.ts
