+++ scripts/poolConfigs/types.ts
