+++ scripts/poolConfigs/utils/prefetchTokenSymbols.ts
