+++ scripts/printPoolConfigs.ts
@@ -4,4 +4,4 @@
  * This is a wrapper script that uses the modular implementation.
  * For better maintainability, the implementation has been moved to the poolConfig folder.
  */
-import './poolConfig/index' 
+import './poolConfigs/index'
